+++ pom.xml
@@ -111,7 +111,7 @@
         </dependency>
         <dependency>
             <groupId>com.amazonaws</groupId>
-            <artifactId>aws-java-sdk-ec2</artifactId>
+            <artifactId>aws-java-sdk-ecs</artifactId>
             <version>1.10.51</version>
         </dependency>
         <dependency>
